Basic Sankey Diagram
This commit includes the basic Sankey diagram before I’ve had the
chance to rewrite it to a queuing system. The diagram should be
functional on smaller data sets however it does take some time to run
(often times out on larger data sets).

I will look at optimising the mongo caching and converting this kind of
reporting tool to a request based system where the report can run in
the background.